fixed unit tests and linting issues
